+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [50.1.2](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@50.1.1...@codecademy/styleguide@50.1.2) (2021-11-01)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [50.1.1](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@50.1.0...@codecademy/styleguide@50.1.1) (2021-10-29)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [50.1.0](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@50.0.2...@codecademy/styleguide@50.1.0) (2021-10-28)
+
+
+### Features
+
+* **GlobalPage:** add banner to global page ([44724e8](https://github.com/Codecademy/gamut/commit/44724e80016be619ebc926b8f0ec507e76c7812e))
+
+
+
+### [50.0.2](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@50.0.1...@codecademy/styleguide@50.0.2) (2021-10-28)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [50.0.1](https://github.com/Codecademy/gamut/compare/@codecademy/styleguide@50.0.0...@codecademy/styleguide@50.0.1) (2021-10-27)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
 ## 50.0.0 (2021-10-26)